% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keel.R
\name{create_config}
\alias{create_config}
\title{create config file for use in keel}
\usage{
create_config(train, test)
}
\arguments{
\item{train}{dataset train path}

\item{test}{dataset test path}
}
\description{
create config file for use in keel
}

