% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keel.R
\name{create_dataset}
\alias{create_dataset}
\title{create new dataset for use in keel}
\usage{
create_dataset(dataset, train, label_class)
}
\arguments{
\item{dataset}{dataset to convert in keel format.}

\item{train}{if dataset is train (train=0) or test (train=1).}

\item{label_class}{class attribute of dataset}
}
\value{
file path
}
\description{
create new dataset for use in keel
}

