% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keel.R
\name{insert_attributes}
\alias{insert_attributes}
\title{insert attributes necessary according measured}
\usage{
insert_attributes(pruned, confidence, importance, leaf, metric)
}
\arguments{
\item{pruned}{if pruned tree (pruned=TRUE) or not pruned tree (pruned=FALSE)}

\item{confidence}{value of confidence}

\item{importance}{for MID measure value of importance}

\item{leaf}{number of instances for leaf}

\item{metric}{string value of measure}
}
\description{
insert attributes necessary according measured
}

