% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{monotonicTree}
\alias{monotonicTree}
\title{execute decision tree with metrics.}
\usage{
monotonicTree(data_train = train, data_test = test, metrics = metrics,
  label_class = label_class)
}
\arguments{
\item{data_train}{dataset of train.}

\item{data_test}{dataset of test.}

\item{metrics}{types:MID,RMI....}

\item{label_class}{class of dataset}
}
\description{
execute decision tree with metrics.
}

